fix various compiler errors for clang++ 10.0.1 and macos mojave 10.14.6
Sorry, but I probably broke compilation on Linux with some of the linker flags.

+++ gridnav/scenario.cc
@@ -70,7 +70,7 @@ void Scenario::outputhdr(FILE *out) {
 }
 
 GridMap *Scenario::getmap(std::string mapfile) {
-	std::string path = maproot + mapfile; 
+	std::string path = maproot + mapfile;
 	if (!lastmap || lastmap->file != path) {
 		if (lastmap)
 			delete lastmap;
@@ -89,7 +89,7 @@ Result<GridNav> ScenarioEntry::run(unsigned int n, SearchAlgorithm<GridNav> *src
 	Result<GridNav> &res = srch->res;
 	GridNav::Cost cost = d.pathcost(res.path, res.ops);
 	// Scenario file has 0-cost for no-path.  We use -1.
-	if (fabsf(cost - (double) opt) > Eps && !(opt == 0 && cost == GridNav::Cost(-1)))
+	if (fabs(cost - (double) opt) > Eps && !(opt == 0 && cost == GridNav::Cost(-1)))
 		fatal("Expected optimal cost of %g, got %g\n", opt, (double) cost);
 
 	dfrow(stdout, "run", "uuuuuuuuguugugg",
